Add float cast to ensure getting proper value.

+++ Disks/Disks/src/ofApp.cpp
@@ -29,7 +29,7 @@ void ofApp::update() {
 			{
 				float distance = force_point.distance(disk.pos);
 				ofVec2f distance_vector(force_point - disk.pos);
-				ofVec2f force = (gravity * attractor_mass * disk.mass * distance_vector) / pow(distance * distance + epsilon, 3 / 2);
+				ofVec2f force = (gravity * attractor_mass * disk.mass * distance_vector) / pow(distance * distance + epsilon, float(3 / 2));
 				acceleration += force / disk.mass;
 			}
 		}
